+++ tests/server/blueprints/variants/test_variants_controllers.py
@@ -9,6 +9,8 @@ from scout.server.blueprints.variants.controllers import (
     sv_variants,
 )
 
+from bson.objectid import ObjectId
+
 LOG = logging.getLogger(__name__)
 
 
@@ -39,6 +41,52 @@ def test_gene_panel_choices(institute_obj, case_obj):
 
     # And institute-specific panel should be in the choices as well
     assert ("institute_panel_name", "Institute Panel display name") in panel_options
+
+
+def test_variants_assessment_shared_with_group(real_variant_database, institute_obj, case_obj):
+    # GIVEN a db with variants,
+    adapter = real_variant_database
+    case_id = case_obj["_id"]
+
+    other_case_id = "other_" + case_id
+    other_case_obj = copy.deepcopy(case_obj)
+    other_case_obj["_id"] = other_case_id
+
+    ## WHEN inserting an object with a group id
+    group_id = ObjectId("101010101010101010101010")
+
+    other_case_obj["group"] = [group_id]
+    adapter.case_collection.insert_one(other_case_obj)
+
+    # WHEN setting the same group id for the original case
+    adapter.case_collection.find_one_and_update({"_id": case_id}, {"$set": {"group": [group_id]}})
+
+    # GIVEN a clinical variant from one case
+    variant = adapter.variant_collection.find_one({"case_id": case_id, "variant_type": "clinical"})
+
+    # GIVEN a copy of the variant for the other case
+    other_variant_obj = copy.deepcopy(variant)
+    other_variant_obj["case_id"] = other_case_id
+    other_variant_obj["_id"] = "another_variant"
+    adapter.variant_collection.insert_one(other_variant_obj)
+
+    # WHEN updating an assessment on the same first case variant
+    adapter.variant_collection.find_one_and_update(
+        {"_id": variant["_id"]}, {"$set": {"acmg_classification": 4}}
+    )
+
+    # WHEN retrieving assessments for the variant from the other case
+    variants_query = {"variant_type": "clinical"}
+    variants_query_res = adapter.variants(
+        other_case_id, query=variants_query, category=variant["category"]
+    )
+
+    res = variants(adapter, institute_obj, other_case_obj, variants_query_res, 1000)
+    res_variants = res["variants"]
+
+    # THEN a group assessment is recalled on the other case,
+    # since the variant in the first case had an annotation
+    assert any(variant.get("group_assessments") for variant in res_variants)
 
 
 def test_variants_research_no_shadow_clinical_assessments(