+++ CHANGELOG.md
@@ -16,6 +16,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Bypass igv.js CORS check using `https://github.com/Rob--W/cors-anywhere`
 - Documentation on default and custom IGV.js tracks (admin docs)
 - Lock phenomodels so they're editable by admins only
+- Small case group assessment sharing
 - Support for Font Awesome version 6
 - Submit to Beacon from case page sidebar
 ### Fixed